pyproject.toml
[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.poetry]
name = "geoboard"
version = "0.1.0"
description = ""
authors = ["esutoru"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.88.0"}
pre-commit = "^2.20.0"
mypy = "^0.991"
black = "^22.10.0"
isort = "^5.10.1"
sqlalchemy = {extras = ["mypy"], version = "^1.4.44"}
alembic = "^1.8.1"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
